Added Ajax form on member details page. User details show on same page.
User name autocomplete field added on member details page. When name is submitted, details show on same page. By default user id value is 0.
When Add More Information is clicked a form is opened on popup. When that form is submitted, a new node is created and the member details page updates with the new details.